refactor(workspace): extract YAML playbook types to noetl-executor::playbook (R-1.1 PR-2a)
First of three follow-up PRs that finish R-1.1 of the Rust migration
roadmap (Appendix H of the global hybrid cloud blueprint).  This PR
moves the Pydantic-like YAML playbook types out of playbook_runner.rs
into a new noetl-executor::playbook module so both the CLI's
local-mode runner and the worker's NATS-mode runner (R-1.3) operate
on the same data model.

What this PR moves
------------------

432 lines from src/playbook_runner.rs (lines 13-445 of the previous
revision) → executor/src/playbook.rs:

- RuntimeCapabilities + local() / distributed() factory methods.
- Playbook, Metadata, Executor, ExecutorSpec, ExecutorRequires.
- Step, StepSpec, NextMode, NextRouterSpec, NextArc.
- NextFormat + from_yaml_value() (v10 router AND legacy array
  normalisation).
- ThenBlock, CaseCondition, WhenCondition.
- Tool (all 8 variants: Shell, Http, Playbook, DuckDb, Auth, Sink,
  Rhai, Unsupported).
- AuthConfig, SinkTarget, SinkFormat.
- CmdsList + Default impl.
- NextStep (Canonical / Conditional / NextAction variants).
- LoopConfig.
- Helper functions used in serde defaults:
  default_profile, default_version, default_method, default_duckdb_path.

Visibility widening: types/fields that were `struct`/`enum`/private
inside playbook_runner.rs become `pub struct`/`pub enum`/`pub` in the
executor crate so the impl block in playbook_runner.rs can reach them
across the crate boundary.  No semantic change — these types were
already module-public within the binary; they're now crate-public
within the workspace.

Wiring
------

- repos/cli/Cargo.toml: adds `noetl-executor = { path = "executor" }`
  as a regular (non-dev) dependency.
- executor/src/lib.rs: adds `pub mod playbook;`.
- src/playbook_runner.rs: 432-line types section replaced with an
  11-line `pub use noetl_executor::playbook::{...}` re-export block.
  `#[allow(unused_imports)]` because some helpers are referenced via
  serde-default string attributes that move with their types.  R-1.1
  PR-2b extracts the parser logic and most of these become actively
  used.

What does NOT change
--------------------

- No CLI binary behaviour change.  PlaybookRunner public API,
  RunOutcome envelope, command-line flags, and event-emission shape
  are unchanged.
- main.rs is untouched (all four call sites at lines 2216, 2341,
  6435, 6475 keep using `playbook_runner::PlaybookRunner::new(path)`).
- noetl-tools dependency is not added yet (R-1.1 PR-2c does that
  when it wires the tool dispatch).

Tests
-----

cargo check --workspace: green (only pre-existing dead-code warning
on extract_auth0_tenant_from_domain in main.rs, unrelated).

cargo test --workspace: 3/3 noetl-executor unit tests pass.  Full
CLI test suite would run via the existing pre-merge gate.

Stats
-----

- playbook_runner.rs: 2,688 -> 2,277 lines (-411 net after the
  11-line re-export block).
- executor/src/playbook.rs: 432 new lines.
- Files changed: 5 (3 modified, 1 new module file, Cargo.lock).

Next: R-1.1 PR-2b lifts the parser + command-generation logic
(~1,200 LoC) from playbook_runner.rs into
executor/src/sources/local_playbook.rs.

Refs noetl/cli#19 (partial — closes after PR-2d)
Refs noetl/ai-meta#30
