+++ Cargo.toml
@@ -28,6 +28,13 @@ name = "ntl"
 path = "src/main.rs"
 
 [dependencies]
+# Workspace member crate shipping the shared execution core.  R-1.1
+# PR-2a moves the YAML playbook types out of playbook_runner.rs into
+# noetl-executor::playbook so both the CLI and noetl-worker (R-1.3)
+# operate on the same data model.  See Appendix H of the global hybrid
+# cloud blueprint.
+noetl-executor = { path = "executor" }
+
 clap = { version = "4.5", features = ["derive"] }
 reqwest = { version = "0.12", default-features = false, features = [
   "json",
