+++ src/playbook_runner.rs
@@ -10,439 +10,36 @@ use std::path::PathBuf;
 use std::process::Command;
 use std::sync::{Arc, Mutex};
 
-/// Runtime capability set - defines what features a runtime supports
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RuntimeCapabilities {
-    pub runtime: String, // "local" or "distributed"
-    pub version: String, // "noetl-runtime/1"
-    pub tools: Vec<String>,
-    pub features: Vec<String>,
-}
+// ---------------------------------------------------------------------------
+// YAML playbook types (R-1.1 PR-2a, see Appendix H of the global hybrid
+// cloud blueprint).
+// ---------------------------------------------------------------------------
+//
+// The Pydantic-like types that used to live here (Playbook, Executor,
+// Step, NextFormat, Tool, AuthConfig, SinkTarget, CmdsList, NextStep,
+// LoopConfig, RuntimeCapabilities, and their impls) moved to
+// `noetl-executor::playbook` so the same data model serves both the
+// CLI's local-mode runner and the worker's NATS-mode runner (R-1.3).
+//
+// Field accessors and trait derives are preserved verbatim; the only
+// difference is that types/fields that were private to this file are
+// now `pub` in the executor crate so the impl below can reach them
+// across the crate boundary.  No behaviour change.
+//
+// `#[allow(unused_imports)]` because some types (helpers used via
+// `#[serde(default = "..")]` attributes inside the executor crate, and
+// types not referenced by name in this file's impl block) get pulled in
+// for completeness.  R-1.1 PR-2b extracts the parser logic and most of
+// these become actively used; for now the allow keeps the re-export
+// glob clean and the impl below stable.
+#[allow(unused_imports)]
+pub use noetl_executor::playbook::{
+    default_duckdb_path, default_method, default_profile, default_version, AuthConfig,
+    CaseCondition, CmdsList, Executor, ExecutorRequires, ExecutorSpec, LoopConfig, Metadata,
+    NextArc, NextFormat, NextMode, NextRouterSpec, NextStep, Playbook, RuntimeCapabilities,
+    SinkFormat, SinkTarget, Step, StepSpec, ThenBlock, Tool, WhenCondition,
+};
 
-impl RuntimeCapabilities {
-    /// Local runtime capabilities
-    pub fn local() -> Self {
-        Self {
-            runtime: "local".to_string(),
-            version: "noetl-runtime/1".to_string(),
-            tools: vec![
-                "shell".to_string(),
-                "http".to_string(),
-                "duckdb".to_string(),
-                "rhai".to_string(),
-                "playbook".to_string(),
-                "auth".to_string(),
-                "sink".to_string(),
-            ],
-            features: vec![
-                "case_v1".to_string(),
-                "case_v2".to_string(), // Rhai conditions
-                "loop_v1".to_string(),
-                "vars_v1".to_string(),
-                "jinja2".to_string(),
-            ],
-        }
-    }
-
-    /// Distributed runtime capabilities
-    #[allow(dead_code)]
-    pub fn distributed() -> Self {
-        Self {
-            runtime: "distributed".to_string(),
-            version: "noetl-runtime/1".to_string(),
-            tools: vec![
-                "shell".to_string(),
-                "http".to_string(),
-                "postgres".to_string(),
-                "duckdb".to_string(),
-                "python".to_string(),
-                "playbook".to_string(),
-                "iterator".to_string(),
-            ],
-            features: vec![
-                "case_v1".to_string(),
-                "case_v2".to_string(),
-                "loop_v1".to_string(),
-                "loop_v2".to_string(), // Pagination
-                "vars_v1".to_string(),
-                "vars_v2".to_string(), // Cross-step results
-                "sink_v1".to_string(),
-                "jinja2".to_string(),
-                "event_sourcing".to_string(),
-            ],
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub struct Playbook {
-    #[serde(rename = "apiVersion")]
-    api_version: String,
-    #[allow(dead_code)]
-    kind: String,
-    metadata: Metadata,
-    /// Runtime requirements and capabilities (8-char root key)
-    #[serde(default)]
-    executor: Option<Executor>,
-    workload: Option<HashMap<String, serde_yaml::Value>>,
-    workflow: Vec<Step>,
-}
-
-/// Executor specification - runtime requirements and capabilities
-#[derive(Debug, Deserialize, Default)]
-pub struct Executor {
-    /// Runtime profile: local, distributed, or auto
-    #[serde(default = "default_profile")]
-    pub profile: String,
-    /// Semantic contract version: noetl-runtime/1
-    #[serde(default = "default_version")]
-    pub version: String,
-    /// Required capabilities
-    #[serde(default)]
-    pub requires: Option<ExecutorRequires>,
-    /// Executor spec for entry/final step configuration
-    #[serde(default)]
-    pub spec: Option<ExecutorSpec>,
-}
-
-/// Executor spec for workflow entry and termination control
-#[derive(Debug, Deserialize, Default, Clone)]
-pub struct ExecutorSpec {
-    /// Override entry step (default: workflow[0])
-    #[serde(default)]
-    pub entry_step: Option<String>,
-    /// Optional finalization step run after quiescence
-    #[serde(default)]
-    pub final_step: Option<String>,
-    /// Treat "no next match" as error (default: false = branch terminates)
-    #[serde(default)]
-    pub no_next_is_error: Option<bool>,
-}
-
-fn default_profile() -> String {
-    "auto".to_string()
-}
-fn default_version() -> String {
-    "noetl-runtime/1".to_string()
-}
-
-/// Executor requirements
-#[derive(Debug, Deserialize, Default)]
-pub struct ExecutorRequires {
-    /// Required tool kinds
-    #[serde(default)]
-    pub tools: Vec<String>,
-    /// Required features
-    #[serde(default)]
-    pub features: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct Metadata {
-    name: String,
-    #[allow(dead_code)]
-    path: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct Step {
-    step: String,
-    desc: Option<String>,
-    /// Step enablement guard - evaluated before step runs (canonical v2)
-    #[serde(rename = "when")]
-    when_guard: Option<String>,
-    /// Step-level input data for cross-boundary propagation (DSL v2)
-    #[serde(default)]
-    input: Option<HashMap<String, serde_yaml::Value>>,
-    tool: Option<Tool>,
-    /// Next transitions - raw YAML, parsed manually to support both v10 router and legacy formats
-    #[serde(default)]
-    next: Option<serde_yaml::Value>,
-    #[serde(rename = "case")]
-    case: Option<Vec<CaseCondition>>,
-    #[serde(rename = "loop")]
-    #[allow(dead_code)]
-    loop_config: Option<LoopConfig>,
-    vars: Option<HashMap<String, String>>,
-    /// Step spec for routing mode
-    #[serde(default)]
-    spec: Option<StepSpec>,
-}
-
-/// Step specification for routing control
-#[derive(Debug, Deserialize, Default, Clone)]
-struct StepSpec {
-    /// Routing mode: exclusive (default, first match) or inclusive (all matches)
-    #[serde(default)]
-    pub next_mode: Option<NextMode>,
-}
-
-/// Next routing mode
-#[derive(Debug, Deserialize, Clone, Default)]
-#[serde(rename_all = "lowercase")]
-enum NextMode {
-    #[default]
-    Exclusive,
-    Inclusive,
-}
-
-/// V10 router spec for next transitions
-#[derive(Debug, Clone)]
-struct NextRouterSpec {
-    mode: Option<String>,
-}
-
-/// V10 arc for router format
-#[derive(Debug, Clone)]
-struct NextArc {
-    step: String,
-    when_condition: Option<String>,
-    args: Option<HashMap<String, serde_yaml::Value>>,
-}
-
-/// Next format - supports both v10 router and legacy array formats
-#[derive(Debug)]
-enum NextFormat {
-    /// V10 router format: { spec: { mode: ... }, arcs: [...] }
-    Router {
-        spec: Option<NextRouterSpec>,
-        arcs: Vec<NextArc>,
-    },
-    /// Legacy array format: [{ step: ... }, ...]
-    Array(Vec<NextStep>),
-}
-
-impl NextFormat {
-    /// Parse next field from serde_yaml::Value
-    fn from_yaml_value(value: &serde_yaml::Value) -> Option<NextFormat> {
-        match value {
-            serde_yaml::Value::Sequence(_arr) => {
-                // Legacy array format
-                let steps: Vec<NextStep> = serde_yaml::from_value(value.clone()).ok()?;
-                Some(NextFormat::Array(steps))
-            }
-            serde_yaml::Value::Mapping(map) => {
-                // V10 router format: { spec: { mode: ... }, arcs: [...] }
-                let spec = map.get(&serde_yaml::Value::String("spec".to_string())).and_then(|v| {
-                    if let serde_yaml::Value::Mapping(spec_map) = v {
-                        let mode = spec_map
-                            .get(&serde_yaml::Value::String("mode".to_string()))
-                            .and_then(|m| m.as_str().map(|s| s.to_string()));
-                        Some(NextRouterSpec { mode })
-                    } else {
-                        None
-                    }
-                });
-
-                let arcs = map.get(&serde_yaml::Value::String("arcs".to_string())).and_then(|v| {
-                    if let serde_yaml::Value::Sequence(arcs_arr) = v {
-                        let arcs: Vec<NextArc> = arcs_arr
-                            .iter()
-                            .filter_map(|arc_val| {
-                                if let serde_yaml::Value::Mapping(arc_map) = arc_val {
-                                    let step = arc_map
-                                        .get(&serde_yaml::Value::String("step".to_string()))
-                                        .and_then(|s| s.as_str().map(|s| s.to_string()))?;
-                                    let when_condition = arc_map
-                                        .get(&serde_yaml::Value::String("when".to_string()))
-                                        .and_then(|w| w.as_str().map(|s| s.to_string()));
-                                    let args = arc_map
-                                        .get(&serde_yaml::Value::String("args".to_string()))
-                                        .and_then(|a| serde_yaml::from_value(a.clone()).ok());
-                                    Some(NextArc {
-                                        step,
-                                        when_condition,
-                                        args,
-                                    })
-                                } else {
-                                    None
-                                }
-                            })
-                            .collect();
-                        Some(arcs)
-                    } else {
-                        None
-                    }
-                })?;
-
-                Some(NextFormat::Router { spec, arcs })
-            }
-            _ => None,
-        }
-    }
-}
-
-/// Then block can be either a list of actions or a single action dict.
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum ThenBlock {
-    /// Single action object (backwards compatible)
-    Single(serde_yaml::Value),
-    /// List of action objects
-    List(Vec<NextStep>),
-}
-
-#[derive(Debug, Deserialize)]
-struct CaseCondition {
-    #[serde(flatten)]
-    when: WhenCondition,
-    then: ThenBlock,
-    #[serde(rename = "else")]
-    else_steps: Option<Vec<NextStep>>,
-}
-
-/// Condition that can be either a simple template string or Rhai code
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum WhenCondition {
-    /// Rhai expression for complex conditions
-    Rhai {
-        #[serde(alias = "when_rhai")]
-        rhai: String,
-    },
-    /// Simple template string condition (Jinja2-style)
-    Simple { when: String },
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(tag = "kind", rename_all = "lowercase")]
-enum Tool {
-    Shell {
-        #[serde(default)]
-        cmds: CmdsList,
-    },
-    Http {
-        #[serde(default = "default_method")]
-        method: String,
-        url: String,
-        #[serde(default)]
-        headers: HashMap<String, String>,
-        #[serde(default)]
-        params: HashMap<String, String>,
-        body: Option<String>,
-        #[serde(default)]
-        auth: Option<AuthConfig>,
-    },
-    Playbook {
-        path: String,
-        /// Legacy args field (DSL v1) - deprecated in favor of input
-        #[serde(default)]
-        args: HashMap<String, String>,
-        /// Canonical input field (DSL v2) - takes precedence over args
-        #[serde(default)]
-        input: HashMap<String, serde_yaml::Value>,
-    },
-    #[serde(rename = "duckdb")]
-    DuckDb {
-        #[serde(default = "default_duckdb_path")]
-        db: String,
-        query: Option<String>,
-        #[serde(default)]
-        params: Vec<String>,
-    },
-    Auth {
-        provider: String,
-        #[serde(default)]
-        scopes: Vec<String>,
-        #[serde(default)]
-        project: Option<String>,
-    },
-    Sink {
-        target: SinkTarget,
-        #[serde(default)]
-        format: SinkFormat,
-    },
-    Rhai {
-        code: String,
-        #[serde(default)]
-        args: HashMap<String, String>,
-    },
-    #[serde(other)]
-    Unsupported,
-}
-
-fn default_method() -> String {
-    "GET".to_string()
-}
-
-fn default_duckdb_path() -> String {
-    ".noetl/state.duckdb".to_string()
-}
-
-#[derive(Debug, Deserialize, Clone)]
-struct AuthConfig {
-    /// Auth provider type: adc (Application Default Credentials), token, basic
-    #[serde(alias = "source")]
-    provider: String,
-    #[serde(default)]
-    scopes: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(tag = "type", rename_all = "lowercase")]
-enum SinkTarget {
-    File {
-        path: String,
-    },
-    #[serde(rename = "duckdb")]
-    DuckDb {
-        db: String,
-        table: String,
-    },
-    Gcs {
-        bucket: String,
-        path: String,
-    },
-}
-
-#[derive(Debug, Deserialize, Default)]
-#[serde(rename_all = "lowercase")]
-enum SinkFormat {
-    #[default]
-    Json,
-    Yaml,
-    Csv,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum CmdsList {
-    Single(String),
-    Multiple(Vec<String>),
-}
-
-impl Default for CmdsList {
-    fn default() -> Self {
-        CmdsList::Multiple(vec![])
-    }
-}
-
-/// Next step definition - supports canonical v2 format
-#[derive(Debug, Deserialize)]
-#[serde(untagged)]
-enum NextStep {
-    /// Canonical v2 format: { step: "name", when: "condition", args: {...} }
-    Canonical {
-        step: String,
-        #[serde(rename = "when")]
-        when_condition: Option<String>,
-        #[serde(default)]
-        args: Option<HashMap<String, serde_yaml::Value>>,
-    },
-    /// Legacy conditional: { when: "condition", then: [...] }
-    Conditional { when: Option<String>, then: Vec<NextStep> },
-    /// Legacy next action: { next: [...] }
-    NextAction { next: Vec<NextStep> },
-}
-
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-struct LoopConfig {
-    #[serde(rename = "in")]
-    in_collection: String,
-    iterator: String,
-    mode: Option<String>,
-}
 
 /// Structured outcome of a local playbook run.
 ///
